requirements.txt
# Core dependencies for macOS integration
pyobjc-core>=10.0
pyobjc-framework-EventKit>=10.0
pyobjc-framework-Cocoa>=10.0

# LLM integration
openai>=1.0.0
anthropic>=0.7.0

# Date/time parsing and handling
python-dateutil>=2.8.0
dateparser>=1.1.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0

# Optional: Local LLM support
# ollama>=0.1.0
